+++ src/pages/auth/SignIn.tsx
@@ -1,16 +1,13 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import * as yup from 'yup';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { useNavigate } from 'react-router-dom';
 import { useSignInMutation } from '../../store/services/authApi';
-import { useAppSelector } from '../../hooks/rtk-hooks';
-import { RootState } from '../../store';
 import { useGoogleRecaptcha } from '../../hooks/useGoogleRecaptcha';
 import Button from '../../widgets/Button';
 import { GoSync } from 'react-icons/go';
 import { MdLogin } from 'react-icons/md';
-import getLoginState from '../../utils/getLoginState';
 
 export interface SignInForm {
   username: string;
@@ -32,10 +29,6 @@ const schema = yup
 
 const SignIn = () => {
   // 1. 리덕스 스토어 객체 가져오기
-  const { rememberMe, accessToken, validUntil } = useAppSelector(
-    (state: RootState) => state.auth
-  );
-
   // 2. 리액트 라우터 네비게이션 객체 가져오기
   const navigate = useNavigate();
 
@@ -83,14 +76,6 @@ const SignIn = () => {
   };
 
   // 7. useEffect
-  useEffect(() => {
-    if (
-      getLoginState(rememberMe, accessToken, validUntil) === 'AUTHENTICATED'
-    ) {
-      navigate('/');
-    }
-  }, [rememberMe, accessToken, validUntil]);
-
   // 8. 이벤트 핸들러
 
   // 9. JSX 반환
